requirements.txt
neodroid-flatbuffers
draugr
warg
pyfiglet
fire>=0.1.3
pyzmq>=18.0.1
numpy>=1.16.3
Pillow>=6.0.0
gym>=0.12.0
tqdm>=4.31.1
matplotlib>=3.0.3
cloudpickle>=0.8.1
